% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_moments.R
\name{aggregate_moments}
\alias{aggregate_moments}
\title{Aggregation of stationary or predictive moments}
\usage{
aggregate_moments(momentsObj, aggregation_matrix, by_timepoint = FALSE)
}
\arguments{
\item{momentsObj}{an object of class \code{moments_hhh4} containing stationary or predictive moments, as returned by \code{stationary_moments} or \code{predictive_moments}}

\item{aggregation_matrix}{an aggregation matrix with either \code{momentsObj$n_units} columns
(for aggregation across units while keeping the temporal structure; set option \code{by_timepoint = TRUE} in this case) 
or \code{length(momentsObj$mu_vector)} (for aggregation that does not preserve the temporal structure; set option \code{by_timepoint = FALSE}).}

\item{by_timepoint}{Is aggregation only across units while preserving the temporal structure?}
}
\description{
Aggregation of stationary or predictive moments as calculated from \code{hhh4} objects using \code{stationary_moments} or \code{predictive_moments}
}
