% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_oneStepAhead.R
\name{confint.oneStepAhead}
\alias{confint.oneStepAhead}
\title{confidence intervals for one-step-ahead predictions}
\usage{
\method{confint}{oneStepAhead}(object, parm, level = 0.95, ...)
}
\description{
confidence intervals for one-step-ahead predictions
}
