% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{decompose.hhh4}
\alias{decompose.hhh4}
\title{A wrapper around \code{decompose.hhh4lag} and \code{surveillance::decompose.hhh4}}
\usage{
decompose.hhh4(x, coefs = x$coefficients, ...)
}
\description{
A wrapper around \code{decompose.hhh4lag} and \code{surveillance::decompose.hhh4}
to handle ordinary \code{hhh4} objects and objects of the new \code{hhh4lag} class.
}
