% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{decompose.hhh4lag}
\alias{decompose.hhh4lag}
\title{A modified version of \code{decompose.hhh4}}
\usage{
decompose.hhh4lag(x, coefs = x$coefficients, ...)
}
\description{
A modified version of \code{decompose.hhh4} to deal with the added
features of the \code{hhh4lag} class.
}
