% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{dengueSJ}
\alias{dengueSJ}
\title{Data set on dengue in San Juan, Puerto Rico}
\source{
Counts retrieved from the supplement of Ray et al (2017): Infectious disease prediction with kernel conditional density estimation, Statistics in Medicine 36(30):4908-4929.
These data originally stem from a forecasting competition organized by the US federal government: http://dengueforecasting.noaa.gov/
}
\description{
Case counts of dengue in San Juan, Puerto Rico, 1990-2013; stored as an sts object
}
\author{
Johannes Bracher
}
\keyword{data}
