% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_lags.R
\name{discrete_gamma_lag}
\alias{discrete_gamma_lag}
\title{This function generates (shifted) discrete gamma weights which are subsequently used inside of \code{get_weighted_lags}. To be passed
to \code{hhh4_lag} or \code{profile_par_lag} as the \code{control$funct_lag} argument.}
\usage{
discrete_gamma_lag(par_lag, min_lag, max_lag)
}
\arguments{
\item{par_lag}{a parameter vector of length 2 to steer the lag structure, here \eqn{log(shape)} and \eqn{log(rate)},
where \eqn{shape} and \eqn{rate} are the parameters of the discrete gamma distribution as implemented in the \code{extraDistr} package.}

\item{min_lag}{smallest lag to include; the support of the Poisson form starts only at \code{min_lag}. Defaults to 1.}

\item{max_lag}{highest lag to include; higher lags are cut off and he remaining weights standardized. Defaults to 5.}
}
\description{
This function generates (shifted) discrete gamma weights which are subsequently used inside of \code{get_weighted_lags}. To be passed
to \code{hhh4_lag} or \code{profile_par_lag} as the \code{control$funct_lag} argument.
}
\author{
Maria Dunbar, Johannes Bracher
}
