% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{distr_lag}
\alias{distr_lag}
\title{Display the function and parameters used for distributed lags}
\usage{
distr_lag(hhh4Obj)
}
\arguments{
\item{hhh4Obj}{The hhh4 object for which the parameters used for distributed lags should be displayed.}
}
\description{
Display the function and parameters used for distributed lags
}
