% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanplot_prediction.R
\name{fanplot_prediction}
\alias{fanplot_prediction}
\title{Display prediction as a fan plot}
\usage{
fanplot_prediction(pred, unit, probs = 1:99/100, interpolate_probs = TRUE,
  add_observed = TRUE, fan.col = colorRampPalette(c("darkgreen", "gray90")),
  pt.col = "red", pt.cex = 0.3, l.col = "black", ln = NULL,
  rlab = NULL, add = FALSE, add_legend = FALSE, width_legend = 0.1 *
  (max(pred$timepoints) - min(pred$timepoints))/pred$frequency,
  probs_legend = c(1, 5, 10, 15, 19)/20, ylim = NULL, xlab = "t",
  ylab = "cases", return_matrix = FALSE, ...)
}
\arguments{
\item{pred}{the prediction as returned by \code{longterm_prediction_hhh4}
(and potentially aggregated using aggregate_prediction)}

\item{unit}{numeric denoting the unit to display}

\item{probs}{vector of probabilities: which quantiles shall be displayed in the fan plot?}

\item{interpolate_probs}{logical: smooth curves by simple interpolation of quantiles}

\item{add_observed}{logical: shall observed values be added?}

\item{fan.col, ln, rlab}{graphical parameters passed on to \code{fanplot::fan}}

\item{pt.col, pt.cex, l.col}{graphical parameters for display of observed values}

\item{add}{logical: add to existing plot?}

\item{add_legend}{logical: shall a color key legend be added?}

\item{width_legend}{width of box for color key legend in user coordinates}

\item{probs_legend}{vecor of probabilities to display in the legend}

\item{ylim}{limit for the y-axis, passed to \code{plot()}}

\item{xlab, ylab}{axis labels}

\item{return_matrix}{logical: return matrix passed to \code{fanplot::fan}; useful to make more sophisticated plots.}

\item{...}{other arguments passed on to \code{plot()}}
}
\description{
Plots a fanplot to display quantiles of (negative binomial approximations) of the week-wise predictive distributions
}
