% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4_new.R
\name{fit_par_lag}
\alias{fit_par_lag}
\title{Estimating the lag decay parameter of an \code{hhh4_lag} model using profile likelihood}
\usage{
fit_par_lag(stsObj, control, check.analyticals = FALSE, range_par,
  use_update = TRUE)
}
\arguments{
\item{range_par}{a vector of values to try for the \code{par_lag} argument of \code{funct_lag}}

\item{use_update}{should results from previous values in range_par be used as starting value for next iteration (via \code{update})?}
}
\description{
Wrapper around \code{hhh4_lag} to allow for profile likelihood estimation of the scalar parameter
governing the lag structure. \code{hhh4_lag} can fit models with fixed lag decay parameter; \code{fit_par_lag} loops
around it and tries a set of possible parameters provided in the argument \code{range_par}. NOTE: this will
soon be replaced by \code{profile_par_lag} which does the same, but using \code{optim..., method = "Brent", ...)}.
}
