% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_diags_of_array}
\alias{get_diags_of_array}
\title{Get diagnoal elements of all slices of an array}
\usage{
get_diags_of_array(arr)
}
\arguments{
\item{arr}{An array.}
}
\description{
Extracts diagonals of all slices of an array (i.e. of \code{arr[,,1], arr[,,2], ...} and stacks them in one vector.)
}
