% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{get_weighted_lags}
\alias{get_weighted_lags}
\title{Transform matrix of first-order lagged observations to matrix of weighted sums of past observation}
\usage{
get_weighted_lags(lag1, lag_weights, sum_up = FALSE)
}
\arguments{
\item{lag1}{a matrix of first lags as usually used in \code{hhh4}.}

\item{sum_up}{\code{sum_up = FALSE} returns a more detailed output; for debugging only.}

\item{lag_weigths}{a vector of weights; the length of this vector determines the number of lags.}
}
\description{
This function modifies the design matrices from first-order lags to weighted lags with weighted structure. Used within \code{weightedSumNE}
and \code{weightedSumAR}.
}
