% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanplot_prediction.R
\name{interpolate_qnbinom}
\alias{interpolate_qnbinom}
\title{Interpolate between quantiles to avoid edgy display}
\usage{
interpolate_qnbinom(p, ...)
}
\arguments{
\item{p}{A vector of probabilities for which to obtain interpolated quantiles}

\item{...}{other arguments passed on to \code{pnbinom}}
}
\description{
An auxiliary function used in \code{fanplot_prediction}, \code{fanplot_stationary}
}
