% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_tilde.R
\name{is_complex_neighbourhood}
\alias{is_complex_neighbourhood}
\title{Determine whether an hhh4 object was fitted using one of the more complex techniques for handling neighbourhoods}
\usage{
is_complex_neighbourhood(hhh4Obj)
}
\arguments{
\item{hhh4Obj}{an hhh4 object}
}
\description{
Determine whether an hhh4 object was fitted using one of the more complex techniques for handling neighbourhoods
}
