% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_tilde_complex_neighbourhood.R
\name{lambda_tilde_complex_neighbourhood}
\alias{lambda_tilde_complex_neighbourhood}
\title{Extracting Lambda_Tilde from an hhh4 object with complex neighbourhood structure}
\usage{
lambda_tilde_complex_neighbourhood(hhh4Obj, subset = NULL, periodic = FALSE)
}
\arguments{
\item{hhh4Obj}{a hhh4 object for which to extract Lambda_tilde}

\item{subset}{a subset (in time); only required when periodic == FALSE}

\item{periodic}{choose subset to correspond to one full cycle}
}
\description{
Extracting Lambda_Tilde from an hhh4 object with complex neighbourhood structure. Used for calculations of
longterm predictions and stationary distributions.
}
