% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{linear_lag}
\alias{linear_lag}
\title{Function to obtain linearly decaying weights}
\usage{
linear_lag(par_lag, min_lag, max_lag)
}
\arguments{
\item{par_lag}{a parameter to steer the lag structure, here \eqn{logit(m)} where \eqn{m} is the linear decay
factor; see details of \code{hhh4lag} or \code{profile_par_lag}.}

\item{min_lag}{smallest lag to include; the support of the Poisson form starts only at \code{min_lag}. Defaults to 1.}

\item{max_lag}{highest lag to include; higher lags are cut off and he remaining weights standardized. Defaults to 5.}
}
\description{
This function generates linearly decaying weights which are subsequently used inside of \code{get_weighted_lags}. To be passed
to \code{hhh4_lag} or \code{profile_par_lag} as the \code{control$funct_lag} argument. There are different ways of
specifying linearly decaying weights, the version implemented here is
\deqn{u0_q = max(1 - mq, 0)}
and \eqn{u_q = u0_q / sum_{q = 1}^Q u0_q} for \eqn{q = 1, ..., Q}.
}
