% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{logLik.hhh4lag}
\alias{logLik.hhh4lag}
\title{A modified version of \code{logLik.hhh4}}
\usage{
\method{logLik}{hhh4lag}(object, ...)
}
\description{
A modified version of \code{terms.hhh4} to deal with the added
features of the \code{logLik.hhh4} class.
}
