% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{matrix_is_cyclic}
\alias{matrix_is_cyclic}
\title{Check whether the rows of a matrix show a cyclic pattern}
\usage{
matrix_is_cyclic(matr, length_of_period)
}
\arguments{
\item{matr}{The parameter matrix to check.}

\item{length_of_period}{Usually 52 (52 weeks per year).}
}
\value{
logical: does the matrix show a periodic pattern?
}
\description{
Needed to determine whether \code{stationary_moments} is applicable (works only for
models with periodic parameter structure)
}
