% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{neOffsetArray.hhh4lag}
\alias{neOffsetArray.hhh4lag}
\title{A modified version of \code{neOffsetArray}}
\usage{
neOffsetArray.hhh4lag(object, pars = coefW(object),
  subset = object$control$subset)
}
\description{
A modified version of \code{neOffsetArray} to deal with the added
features of the \code{hhh4lag} class.
}
