% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{noroBL}
\alias{noroBL}
\title{Data set on norovirus gastroenteritis in Bremen and Lower Saxony}
\source{
Surveillance counts retrieved from SurvStat@RKI 2.0 service (https://survstat.rki.de), Robert Koch Institute, Berlin as of 30 May 2017.
}
\description{
Case counts of noroviral gastroenteritis in the German states of Bremen and Lower Saxony, 2011-2017; stored as an sts object
}
\author{
Johannes Bracher
}
\keyword{data}
