% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4_new.R
\name{numeric_fisher_hhh4lag}
\alias{numeric_fisher_hhh4lag}
\title{Numerical evaluation of the covariance matrix including the additional parameter
\code{par_lg}}
\usage{
numeric_fisher_hhh4lag(best_mod)
}
\arguments{
\item{best_mod}{an \code{hhh4lag} object; should be generated in \code{profile_par_lag} so that
the \code{par_lag} parameter is already optimized.}
}
\description{
Numerical evaluation of the covariance matrix including the additional parameter
\code{par_lg}
}
