% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_oneStepAhead.R
\name{plot.oneStepAhead}
\alias{plot.oneStepAhead}
\title{simple plot of one-step-ahead forecasts}
\usage{
\method{plot}{oneStepAhead}(x, unit = 1, probs = 1:99/100, start = NULL, means.args = NULL, ...)
}
\description{
simple plot of one-step-ahead forecasts
}
