% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{poisson_lag}
\alias{poisson_lag}
\title{Transform matrix of first lags to matrix of Poisson lags
To be passed to \code{hhh4lag} or \code{profile_par_lag} in order to use Poisson lags.}
\usage{
poisson_lag(lag1, par_lag, min_lag, max_lag, sum_up = TRUE)
}
\arguments{
\item{lag1}{a matrix of first lags as usually used in \code{hhh4}.}

\item{par_lag}{a parameter to steer the lag structure, here \eqn{log(\mu)} where \eqn{\mu} is the parameter of
the Poisson distribution characterizing the lag structure; see details of \code{hhh4lag} or \code{profile_par_lag}}

\item{min_lag}{smallest lag to include; the support of the Poisson form starts only at \code{min_lag}. Defaults to 1.}

\item{max_lag}{highest lag to include; higher lags are cut off and he remaining weights standardized. Defaults to 5.}

\item{sum_up}{\code{sum_up = FALSE} returns a more detailed output; for debugging only}
}
\description{
Transform matrix of first lags to matrix of Poisson lags
To be passed to \code{hhh4lag} or \code{profile_par_lag} in order to use Poisson lags.
}
