% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{print.hhh4lag}
\alias{print.hhh4lag}
\title{A modified version of \code{surveillance::print.hhh4}}
\usage{
\method{print}{hhh4lag}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
A modified version of \code{surveillance::print.hhh4} to deal with the added
features of the \code{hhh4lag} class.
}
