% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{print.summary.hhh4lag}
\alias{print.summary.hhh4lag}
\title{A modified version of \code{print.summary.hhh4}}
\usage{
\method{print}{summary.hhh4lag}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\description{
A modified version of \code{print.summary.hhh4} to deal with the added
features of the \code{hhh4lag} class.
}
