% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{psi2size.hhh4lag}
\alias{psi2size.hhh4lag}
\title{A modified version of \code{psi2size.hhh4}}
\usage{
psi2size.hhh4lag(object, subset = object$control$subset, units = NULL)
}
\description{
A modified version of \code{psi2size.hhh4} to deal with the added
features of the \code{hhh4lag} class. Extracts estimated overdispersion
in dnbinom() parametrization (and as matrix)
}
