% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_oneStepAhead.R
\name{quantile.oneStepAhead}
\alias{quantile.oneStepAhead}
\title{quantiles of the one-step-ahead forecasts}
\usage{
\method{quantile}{oneStepAhead}(x, probs = c(2.5, 10, 50, 90, 97.5)/100, ...)
}
\description{
quantiles of the one-step-ahead forecasts
}
