% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{ranef.hhh4lag}
\alias{ranef.hhh4lag}
\title{A modified version of \code{ranef.hhh4}}
\usage{
\method{ranef}{hhh4lag}(object, tomatrix = FALSE, intercept = FALSE, ...)
}
\description{
A modified version of \code{ranef.hhh4}
}
