% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{residuals.hhh4lag}
\alias{residuals.hhh4lag}
\title{A modified version of \code{residuals.hhh4}}
\usage{
\method{residuals}{hhh4lag}(object, type = c("deviance", "response"), ...)
}
\description{
A modified version of \code{residuals.hhh4} to deal with the added
features of the \code{hhh4lag} class. Computes deviance residuals.
}
