% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_simulate.R
\name{simulate.hhh4lag}
\alias{simulate.hhh4lag}
\title{Simulate "hhh4_lag" Count Time Series}
\usage{
\method{simulate}{hhh4lag}(
  object,
  nsim = 1,
  seed = NULL,
  y.start = NULL,
  subset = 1:nrow(object$stsObj),
  coefs = coef(object),
  components = c("ar", "ne", "end"),
  simplify = nsim > 1,
  ...
)
}
\description{
This function is the equivalent of \code{surveillance::simulate.hhh4} for model fits of class
\code{hhh4lag}, obtained from \code{hhh4_lag} or \code{profile_par_lag}. The arguments are the
same as in \code{surveillance::simulate.hhh4}, the only difference being that \code{y.start}
needs to be a matrix with \code{object$control$max_lag} rows and \code{object$nUnit} columns.
}
\details{
This function is still being tested!!!
}
