% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationary_moments.R
\name{stationary_moments}
\alias{stationary_moments}
\title{Stationary distributions implied by a \code{hhh4}-model}
\usage{
stationary_moments(hhh4Obj, start = 1, n_seasons = 1,
  return_Sigma = FALSE, return_cov_array = FALSE,
  return_mu_decomposed = FALSE, return_M = FALSE, max.iter = 10,
  tolerance = 1e-05)
}
\arguments{
\item{hhh4Obj}{\code{hhh4} object for which to calculate stationary moments}

\item{start}{start of the season}

\item{n_seasons}{number of}

\item{return_Sigma}{logical: return entire variance/covariance matrix of the prediction; can take a lot of storage}

\item{return_cov_array}{logical: return an array containing week-wise covariance matrices}

\item{return_mu_decomposed}{logical: return an array containing a decomposition of stationary means into the three
components \code{endemic}, \code{epi.own} and \code{epi.others}.}

\item{return_M}{logical: return the array M containing un-centered second moments (used internally for calculations)}

\item{max.iter}{maximum number of iterations before iterative algorithm stops}

\item{tolerance}{element-wise maximum tolerance (entering into temination criterion for the iterative calculation)}
}
\description{
Returns the mean vector and covariance matrix of the stationary distribution implied by an hhh4 object.
}
