% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{summary.hhh4lag}
\alias{summary.hhh4lag}
\title{A modified version of \code{surveillance::summary.hhh4}}
\usage{
\method{summary}{hhh4lag}(object, maxEV = FALSE, ...)
}
\description{
A modified version of \code{surveillance::summary.hhh4} to deal with the added
features of the \code{hhh4lag} class.
}
