% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{terms.hhh4lag}
\alias{terms.hhh4lag}
\title{A modified version of \code{terms.hhh4}}
\usage{
\method{terms}{hhh4lag}(x, ...)
}
\description{
A modified version of \code{terms.hhh4} to deal with the added
features of the \code{hhh4lag} class.
}
