% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{unrestricted_lag}
\alias{unrestricted_lag}
\title{Function to obtain unrestricted lags}
\usage{
unrestricted_lag(par_lag, min_lag, max_lag)
}
\arguments{
\item{par_lag}{theparameter vector of length \eqn{Q - 1}}

\item{min_lag}{smallest lag to include; the support of the Poisson form starts only at \code{min_lag}. Defaults to 1.}

\item{max_lag}{highest lag to include; higher lags are cut off and he remaining weights standardized. Defaults to 5.}
}
\description{
This function generates \eqn{Q} lag weights without a parametric constraint. The weights are obtained via
a multinomial logit transformation where the first lag is the reference category.
}
