% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{weibull_lag}
\alias{weibull_lag}
\title{Discrete Weibull lag function}
\usage{
weibull_lag(par_lag, min_lag, max_lag)
}
\arguments{
\item{par_lag}{the parameter vector (must be of length 2)}

\item{min_lag}{smallest lag to include}

\item{max_lag}{highest lag to include}
}
\value{
the discrete Weibull lag distribution
}
\description{
The discrete Weibull distribution has probability mass function \insertCite{NakagawaOsaki1975}{Rdpack}
\eqn{p(d; \kappa, \beta) = \kappa ^ {(d - 1) ^ \beta} - \kappa ^ {d ^ \beta}, d = 1, 2, \ldots}
which reduces to the geometric if \eqn{\beta = 1} (\code{geometric_lag}) and the discrete Raleigh distribution if \eqn{\beta = 2} \insertCite{Roy2004}{Rdpack}
}
\references{
\insertAllCited
}
\seealso{
geometric_lag poisson_lag ar2_lag linear_lag unrestricted_lag
}
