% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_lags.R
\name{ar2_lag}
\alias{ar2_lag}
\title{Function to obtain AR2 weights
This function generates AR2 weights which are subsequently used inside of \code{get_weighted_lags}. To be passed
to \code{hhh4_lag} or \code{profile_par_lag} as the \code{control$funct_lag} argument.}
\usage{
ar2_lag(par_lag, min_lag, max_lag)
}
\arguments{
\item{par_lag}{a parameter to steer the lag structure, here \eqn{logit(p)} where \eqn{p} is the weight of
the first lag; see details of \code{hhh4lag} or \code{profile_par_lag}.}

\item{min_lag}{smallest lag to include; the support of the Poisson form starts only at \code{min_lag}. Defaults to 1.}

\item{max_lag}{highest lag to include; higher lags are cut off and he remaining weights standardized. Defaults to 5.}
}
\description{
Function to obtain AR2 weights
This function generates AR2 weights which are subsequently used inside of \code{get_weighted_lags}. To be passed
to \code{hhh4_lag} or \code{profile_par_lag} as the \code{control$funct_lag} argument.
}
