% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanplot_stationary.R
\name{fanplot_stationary}
\alias{fanplot_stationary}
\title{Display stationary distribution as a fanplot}
\usage{
fanplot_stationary(stat_mom, unit, probs = 1:99/100,
  interpolate_probs = TRUE, fan.col = colorRampPalette(c("darkgreen",
  "gray90")), pt.col = "red", pt.cex = 0.3, l.col = "black", ln = NULL,
  ln.col = "red", rlab = NULL, style = "fan", add = FALSE,
  timepoints = 1:nrow(stat_mom$mu_matrix)/stat_mom$freq, add_legend = FALSE,
  width_legend = 0.1 * (max(timepoints) - min(timepoints)),
  probs_legend = c(1, 25, 50, 75, 99)/100, hlines = NULL, vlines = NULL,
  ylim = NULL, xlab = "t", ylab = "cases", return_matrix = FALSE, ...)
}
\arguments{
\item{stat_mom}{the stationary moments as returned by \code{stationary_moments_hhh4}
(and potentially aggregated using aggregate_prediction)}

\item{unit}{numeric denoting the unit to display}

\item{probs}{vector of probabilities: which quantiles shall be displayed in the fan plot?}

\item{interpolate_probs}{logical: smooth curves by simple interpolation of quantiles}

\item{fan.col, ln, ln.col, rlab, style}{graphical parameters passed on to \code{fanplot::fan}}

\item{pt.col, pt.cex, l.col}{graphical parameters for display of observed values}

\item{add}{logical: add to existing plot?}

\item{timepoints}{vector giving the x-coordinates for the fanplot (generates \code{start} and \code{frequency} for \code{fanplot::fan})}

\item{add_legend}{logical: shall a color key legend be added?}

\item{width_legend}{width of box for color key legend in user coordinates}

\item{probs_legend}{vecor of probabilities to display in the legend}

\item{hlines, vlines}{coordinates for horizontal and vertical grid lines}

\item{ylim}{limit for the y-axis, passed to \code{plot()}}

\item{xlab, ylab}{axis labels}

\item{return_matrix}{logical: return matrix passed to \code{fanplot::fan}; useful to make more sophisticated plots.}

\item{...}{other arguments passed on to \code{plot()}}
}
\description{
Plots a fanplot to display quantiles of (negative binomial approximations) of the week-wise stationary distributions
}
