% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4_new.R
\name{hhh4_lag}
\alias{hhh4_lag}
\title{Fitting hhh4 models with distributed lags}
\usage{
hhh4_lag(stsObj, control = list(ar = list(f = ~-1, offset = 1, lag = 1,
  funct_lag = geometric_lag, par_lag = list(mu = 1), max_lag = 5, use_distr_lag
  = FALSE), ne = list(f = ~-1, offset = 1, lag = 1, funct_lag = geometric_lag,
  par_lag = list(mu = 1), max_lag = 5, use_distr_lag = FALSE, weights =
  neighbourhood(stsObj) == 1, scale = NULL, normalize = FALSE), end = list(f =
  ~1, offset = 1), family = c("Poisson", "NegBin1", "NegBinM"), subset =
  2:nrow(stsObj), optimizer = list(stop = list(tol = 1e-05, niter = 100),
  regression = list(method = "nlminb"),      variance = list(method =
  "nlminb")), verbose = FALSE, start = list(fixed = NULL, random = NULL, sd.corr
  = NULL), data = list(t = stsObj@epoch - min(stsObj@epoch)), keep.terms =
  FALSE), check.analyticals = FALSE)
}
\arguments{
\item{stsObj, control, check.analyticals}{As in \code{surveillance::hhh4}, but \code{control}
allows for some additional arguments (see details)

In this modified version of \code{surveillance::hhh4}, distributed laga can be specified using the following
additional elements in the \code{ar} and \code{ne} parts of \code{control}:
\itemize{
  \item{\code{funct_lag}}{ Function to calculate weights for different lags; defaults to geometric lags (\code{hhh4addon:::geometric_lag}).
  Function needs to take the following parameters:
  \itemize{
  \item{lag1}{ Matrix containing the time-varying \eqn{\phi_{ijt}} which enter into \eqn{\phi_{ijqt} = \phi_{ijt}*w_q}}
  \item{par_lag}{ A scalar parameter or a list of parameters to steer \eqn{w_q}}
  }}
  \item{\code{max_lag}}{ Maximum number of lags after which}
}}
}
\description{
A modified version of \code{surveillance::hhh4} to allow for distributed lags.
}
