% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_oneStepAhead.R
\name{oneStepAhead_hhh4lag}
\alias{oneStepAhead_hhh4lag}
\title{Predictive Model Assessment for hhh4_lag Models}
\usage{
oneStepAhead_hhh4lag(result, tp, type = c("rolling", "first", "final"),
  which.start = c("current", "final"), keep.estimates = FALSE,
  verbose = TRUE, cores = 1)
}
\description{
A version of \code{surveillance::oneStepAhead} for \code{hhh4_lag} objects. NOTE: the \code{lag_par} parameter
is not re-fit in this function!
}
