% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_moments.R
\name{predictive_moments}
\alias{predictive_moments}
\title{#' Analytical computation of predictive moments for an \code{hhh4} model}
\usage{
predictive_moments(hhh4Obj, t_condition, lgt, return_Sigma = FALSE,
  return_cov_array = FALSE, return_mu_decomposed = FALSE,
  return_M = FALSE)
}
\arguments{
\item{t_condition}{the index of the week on which to condition the path forecast}

\item{lgt}{the length of the path forecast, i.e. 52 for forecasting an entire season when using weekly data}

\item{return_Sigma}{logical: should the entire variance-covariance matrix of the forecast be returned? defaults to
\code{FALSE} in order to save storage.}

\item{return_cov_array}{logical: should an array with week-wise covariance matrices be returned?}

\item{return_mu_decomposed}{logical: should an array with the predictive means decomposed into the fidderent components be returned?}

\item{return_M}{logical: should the matrix M containing the predictive un-centered second moments be returned?}
}
\value{
mu_matrix,mu_vector predictive means as a matrix and flattened as a vector

var_matrix matrix of predictive variances

Sigma predictive variance-covariance matrix

cov_array array containing week-wise predictive variance-covariance matrices.
}
\description{
#' This functions calculates the predictive mean vector and covariance matrix for a path forecast from
an \code{hhh4} model.
}
\details{
#' @param hhh4Obj an \code{hhh4} object
}
