% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhh4lag_methods.R
\name{update.hhh4lag}
\alias{update.hhh4lag}
\title{A modified version of \code{update.hhh4}}
\usage{
\method{update}{hhh4lag}(object, ..., S = NULL, subset.upper = NULL,
  use.estimates = object$convergence, evaluate = TRUE)
}
\description{
A modified version of \code{update.hhh4} to deal with the added
features of the \code{hhh4lag} class.
}
